Something

+++ Controllers/CRUD_USER/Deleteuser.controller.js
@@ -1,16 +1,23 @@
 const { UserModel } = require("../../Models/User.model");
+const bcrypt = require("bcrypt");
 
 const DeleteUser = async (req, res) => {
-  let { id } = req.params;
-  try {
-    if (id) {
-      await UserModel.findByIdAndDelete(id);
-      res.status(200).json("User Deleted Successfully");
+  const { id, password } = req.body;
+  let findbyId = UserModel.findById(id);
+  if (id) {
+    let PasswordCheck = await bcrypt.compare(password, findbyId.password);
+    if (PasswordCheck) {
+      try {
+        await UserModel.findByIdAndDelete(id);
+        res.status(200).json("User Deleted Successfully");
+      } catch (error) {
+        res.status(500).json({ message: error.message });
+      }
     } else {
-      res.status(403).json("You are not Authorised");
+      res.send({ message: "Password Doesn't match." });
     }
-  } catch (error) {
-    res.status(500).json({ message: error.message });
+  } else {
+    res.send({ message: "User Doesn't exist" });
   }
 };
